Version 1.15.1
o Added '...' to readCodelinkSet() so that it pass additional arguments (e.g. 'flag') to readCodelink(). Thank you to Luca Beltrame for the suggestion and patch.
o Now when readCodelinkSet() reads a list of files and the phenoData is not supplied it will create a default one, instead of failing.

Version: 1.15.0

o Version number bumped for BioC devel branch.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/codelink@42947 bc3139a8-67e5-0310-9ffc-ced21a209358